# docs/openapi.yaml
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    NightscoutDayStatusResponse:
      description: Nightscout sync status for one diary day.
      properties:
        accepted_meals_count:
          title: Accepted Meals Count
          type: integer
        configured:
          title: Configured
          type: boolean
        connected:
          title: Connected
          type: boolean
        date:
          format: date
          title: Date
          type: string
        failed_meals_count:
          title: Failed Meals Count
          type: integer
        last_sync_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Last Sync At
        synced_meals_count:
          title: Synced Meals Count
          type: integer
        unsynced_meals_count:
          title: Unsynced Meals Count
          type: integer
      required:
      - date
      - connected
      - configured
      - accepted_meals_count
      - unsynced_meals_count
      - synced_meals_count
      - failed_meals_count
      title: NightscoutDayStatusResponse
      type: object
    NightscoutEventsResponse:
      description: Combined read-only Nightscout context events.
      properties:
        glucose:
          items:
            $ref: '#/components/schemas/NightscoutGlucoseEntryResponse'
          title: Glucose
          type: array
        insulin:
          items:
            $ref: '#/components/schemas/NightscoutInsulinEventResponse'
          title: Insulin
          type: array
      required:
      - glucose
      - insulin
      title: NightscoutEventsResponse
      type: object
    NightscoutGlucoseEntryResponse:
      description: Read-only Nightscout glucose entry normalized for UI context.
      properties:
        source:
          anyOf:
          - type: string
          - type: 'null'
          title: Source
        timestamp:
          format: date-time
          title: Timestamp
          type: string
        trend:
          anyOf:
          - type: string
          - type: 'null'
          title: Trend
        unit:
          default: mmol/L
          title: Unit
          type: string
        value:
          title: Value
          type: number
      required:
      - timestamp
      - value
      title: NightscoutGlucoseEntryResponse
      type: object
    NightscoutImportRequest:
      description: Import Nightscout context into the local read-only cache.
      properties:
        from_datetime:
          format: date-time
          title: From Datetime
          type: string
        import_insulin_events:
          default: true
          title: Import Insulin Events
          type: boolean
        sync_glucose:
          default: true
          title: Sync Glucose
          type: boolean
        to_datetime:
          format: date-time
          title: To Datetime
          type: string
      required:
      - from_datetime
      - to_datetime
      title: NightscoutImportRequest
      type: object
    NightscoutImportResponse:
      description: Nightscout local cache import summary.
      properties:
        from_datetime:
          format: date-time
          title: From Datetime
          type: string
        glucose_imported:
          title: Glucose Imported
          type: integer
        glucose_total:
          title: Glucose Total
          type: integer
        insulin_imported:
          title: Insulin Imported
          type: integer
        insulin_total:
          title: Insulin Total
          type: integer
        last_error:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Error
        to_datetime:
          format: date-time
          title: To Datetime
          type: string
      required:
      - from_datetime
      - to_datetime
      - glucose_imported
      - insulin_imported
      - glucose_total
      - insulin_total
      title: NightscoutImportResponse
      type: object
    NightscoutInsulinEventResponse:
      description: Read-only Nightscout insulin event.
      properties:
        enteredBy:
          anyOf:
          - type: string
          - type: 'null'
          title: Enteredby
        eventType:
          anyOf:
          - type: string
          - type: 'null'
          title: Eventtype
        insulin_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Insulin Type
        insulin_units:
          anyOf:
          - type: number
          - type: 'null'
          title: Insulin Units
        nightscout_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Nightscout Id
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        timestamp:
          format: date-time
          title: Timestamp
          type: string
      required:
      - timestamp
      title: NightscoutInsulinEventResponse
      type: object
...
    NightscoutStatusResponse:
      description: Nightscout optional integration status.
      properties:
        configured:
          title: Configured
          type: boolean
        status:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Status
      required:
      - configured
      title: NightscoutStatusResponse
      type: object
    NightscoutSyncResponse:
      description: Nightscout meal sync response.
      properties:
        nightscout_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Nightscout Id
        nightscout_sync_error:
          anyOf:
          - type: string
          - type: 'null'
          title: Nightscout Sync Error
        nightscout_sync_status:
          anyOf:
          - type: string
          - type: 'null'
          title: Nightscout Sync Status
        nightscout_synced_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Nightscout Synced At
        response:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Response
        synced:
          title: Synced
          type: boolean
      required:
      - synced
      title: NightscoutSyncResponse
      type: object
...
  securitySchemes:
    BearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
info:
  summary: Informational food diary API.
  title: Glucotracker API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /meals/{meal_id}/sync_nightscout:
    post:
      description: Sync an accepted meal as a diary-only Nightscout treatment.
      operationId: syncMealToNightscout
      parameters:
      - in: path
        name: meal_id
        required: true
        schema:
          format: uuid
          title: Meal Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NightscoutSyncResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Sync Meal To Nightscout
      tags:
      - nightscout
      x-glucotracker-required-feature: nightscout
      x-glucotracker-required-role: gluco
      x-glucotracker-scoped: true
  /meals/{meal_id}/unsync_nightscout:
    post:
      description: Delete a remote Nightscout treatment and clear local sync fields.
      operationId: unsyncMealFromNightscout
      parameters:
      - in: path
        name: meal_id
        required: true
        schema:
          format: uuid
          title: Meal Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NightscoutSyncResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Unsync Meal From Nightscout
      tags:
      - nightscout
      x-glucotracker-required-feature: nightscout
      x-glucotracker-required-role: gluco
      x-glucotracker-scoped: true
  /nightscout/day_status:
    get:
      description: Return Nightscout manual-sync counters for a selected day.
      operationId: getNightscoutDayStatus
      parameters:
      - in: query
        name: date
        required: true
        schema:
          format: date
          title: Date
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NightscoutDayStatusResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Get Nightscout Day Status
      tags:
      - nightscout
      x-glucotracker-required-feature: nightscout
      x-glucotracker-required-role: gluco
      x-glucotracker-scoped: true
  /nightscout/events:
    get:
      description: Return combined read-only Nightscout glucose and insulin context
        events.
      operationId: getNightscoutEvents
      parameters:
      - in: query
        name: from
        required: true
        schema:
          format: date-time
          title: From
          type: string
      - in: query
        name: to
        required: true
        schema:
          format: date-time
          title: To
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NightscoutEventsResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Get Nightscout Events
      tags:
      - nightscout
      x-glucotracker-required-feature: nightscout
      x-glucotracker-required-role: gluco
      x-glucotracker-scoped: true
  /nightscout/glucose:
    get:
      description: Return read-only Nightscout glucose entries as gentle context.
      operationId: getNightscoutGlucose
      parameters:
      - in: query
        name: from
        required: true
        schema:
          format: date-time
          title: From
          type: string
      - in: query
        name: to
        required: true
        schema:
          format: date-time
          title: To
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/NightscoutGlucoseEntryResponse'
                title: Response Getnightscoutglucose
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Get Nightscout Glucose
      tags:
      - nightscout
      x-glucotracker-required-feature: nightscout
      x-glucotracker-required-role: gluco
      x-glucotracker-scoped: true
  /nightscout/import:
    post:
      description: Fetch Nightscout glucose/insulin context and cache it locally.
      operationId: importNightscoutContext
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NightscoutImportRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NightscoutImportResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Import Nightscout Context
      tags:
      - nightscout
      x-glucotracker-required-feature: nightscout
      x-glucotracker-required-role: gluco
      x-glucotracker-scoped: true
  /nightscout/insulin:
    get:
      description: Return read-only Nightscout insulin events without linking to dosing.
      operationId: getNightscoutInsulin
      parameters:
      - in: query
        name: from
        required: true
        schema:
          format: date-time
          title: From
          type: string
      - in: query
        name: to
        required: true
        schema:
          format: date-time
          title: To
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/NightscoutInsulinEventResponse'
                title: Response Getnightscoutinsulin
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Get Nightscout Insulin
      tags:
      - nightscout
      x-glucotracker-required-feature: nightscout
      x-glucotracker-required-role: gluco
      x-glucotracker-scoped: true
  /nightscout/latest-reading:
    get:
      description: Return the latest glucose reading from the local Nightscout cache.
      operationId: getNightscoutLatestReading
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NightscoutLatestReadingResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Get Nightscout Latest Reading
      tags:
      - nightscout
      x-glucotracker-required-feature: nightscout
      x-glucotracker-required-role: gluco
      x-glucotracker-scoped: true
  /nightscout/status:
    get:
      description: Return optional Nightscout status without breaking local use.
      operationId: getNightscoutStatus
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NightscoutStatusResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Get Nightscout Status
      tags:
      - nightscout
      x-glucotracker-required-feature: nightscout
      x-glucotracker-required-role: gluco
      x-glucotracker-scoped: true
  /nightscout/sync/today:
    post:
      description: Manually send accepted unsynced meals for a selected day.
      operationId: syncTodayToNightscout
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NightscoutSyncTodayRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NightscoutSyncTodayResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Sync Today To Nightscout
      tags:
      - nightscout
      x-glucotracker-required-feature: nightscout
      x-glucotracker-required-role: gluco
      x-glucotracker-scoped: true
...
  /settings/nightscout:
    get:
      description: Return masked server-side Nightscout settings.
      operationId: getNightscoutSettings
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NightscoutSettingsResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Get Nightscout Settings
      tags:
      - nightscout
      x-glucotracker-required-feature: nightscout
      x-glucotracker-required-role: gluco
      x-glucotracker-scoped: true
    put:
      description: Update server-side Nightscout settings. Secret is write-only.
      operationId: updateNightscoutSettings
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NightscoutSettingsPatch'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NightscoutSettingsResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Update Nightscout Settings
      tags:
      - nightscout
      x-glucotracker-required-feature: nightscout
      x-glucotracker-required-role: gluco
      x-glucotracker-scoped: true
  /settings/nightscout/test:
    post:
      description: Test Nightscout connection and persist masked status.
      operationId: testNightscoutConnection
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NightscoutTestResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Test Nightscout Connection
      tags:
      - nightscout
      x-glucotracker-required-feature: nightscout
      x-glucotracker-required-role: gluco
      x-glucotracker-scoped: true
  /timeline:
    get:
      description: Return backend-owned food episodes with local Nightscout context.
      operationId: getTimeline
      parameters:
      - in: query
        name: from
        required: true
        schema:
          format: date-time
          title: From
          type: string
      - in: query
        name: to
        required: true
        schema:
          format: date-time
          title: To
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/TimelineResponse'
                - $ref: '#/components/schemas/TimelineFoodResponse'
                title: Response Gettimeline
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Get Timeline
      tags:
      - nightscout
      x-glucotracker-food-omits:
      - glucose
      - nightscout
      x-glucotracker-role-variant: true
      x-glucotracker-scoped: true